move shared util function between core and similarity panel to construct run name to fiftyone/utilities format.ts and other misc fixes
